fix/elastic-beanstalk config (#12)
