+++ .ebextensions/01_django.config
@@ -3,7 +3,7 @@ option_settings:
     DJANGO_SETTINGS_MODULE: "evaluation_registry.settings"
     PYTHONPATH: "/var/app/current:$PYTHONPATH"
   aws:elasticbeanstalk:container:python:
-    WSGIPath: "ask_ai.wsgi:application"
+    WSGIPath: "evaluation_registry.wsgi:application"
 container_commands:
   02_migrate:
     command: "source /var/app/venv/*/bin/activate && python3 manage.py migrate --noinput"
