--- Procfile
@@ -1,1 +1,0 @@
-web: python manage.py migrate && waitress-serve --port=$PORT evaluation_registry.wsgi:application
